requirements.txt
protobuf
PyPubSub
paho-mqtt
requests
